display active model below parameters, dont display long description
