+++ application/ui/src/features/models/train-model/model-architectures-list/all-model-architectures.component.tsx
@@ -8,43 +8,8 @@ import { Flex } from '@geti/ui';
 import type { ModelArchitecture as ModelArchitectureType } from '../../../../constants/shared-types';
 import { SortModelArchitectures } from '../sort-model-architectures/sort-model-architectures.component';
 import { SORT_OPTIONS, SORTING_HANDLERS, SortingOptions } from '../sort-model-architectures/utils';
-import { ModelArchitectureCard } from './model-architecture/model-architecture.component';
+import { ModelArchitecture } from './model-architecture.component';
 import { ModelArchitecturesListLayout } from './model-architectures-list-layout/model-architectures-list-layout.component';
-
-type ModelArchitectureProps = {
-    activeModelArchitectureId: string | undefined;
-    modelArchitecture: ModelArchitectureType;
-    selectedModelArchitectureId: string | null;
-    onSelectedModelArchitectureIdChange: (modelArchitectureId: string | null) => void;
-};
-
-const ModelArchitecture = ({
-    activeModelArchitectureId,
-    modelArchitecture,
-    onSelectedModelArchitectureIdChange,
-    selectedModelArchitectureId,
-}: ModelArchitectureProps) => {
-    const isSelected = modelArchitecture.id === selectedModelArchitectureId;
-    const isActive = activeModelArchitectureId === modelArchitecture.id;
-
-    return (
-        <ModelArchitectureCard
-            isCompact
-            modelArchitecture={modelArchitecture}
-            isSelected={isSelected}
-            onSelect={() => onSelectedModelArchitectureIdChange(modelArchitecture.id)}
-        >
-            <Flex direction={'column'} width={'100%'} minWidth={0} gap={'size-100'}>
-                {isActive && <ModelArchitectureCard.Active />}
-                <Flex alignItems={'center'} justifyContent={'space-between'}>
-                    <ModelArchitectureCard.Name />
-                    <ModelArchitectureCard.Description />
-                </Flex>
-            </Flex>
-            <ModelArchitectureCard.Parameters />
-        </ModelArchitectureCard>
-    );
-};
 
 type AllModelArchitecturesProps = {
     activeModelArchitectureId: string | undefined;
